requirements.txt
django==4.2.1
ipython==8.22.2
django-extensions==3.2.3
django-debug-toolbar==4.3.0
django-mptt==0.16.0
Unidecode==1.3.8
Pillow==10.0.0
python-dotenv==1.0.1
django-tailwind==3.8.0
python-slugify==8.0.4
django-ckeditor==6.7.1
psycopg2-binary
enchant==0.0.1
django-tinymce==4.1.0